discord_bot/requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
aiohttp-requests==0.1.2
async-timeout==3.0.1
attrs==19.1.0
cchardet==2.1.4
cffi==1.12.2
chardet==3.0.4
discord.py==1.0.1
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
websockets==6.0
yarl==1.3.0
